Significant Update
We create resource in the ontology with full URI instead of readable
format (prefix:localName format) we used previously.
1 now the user should give the prefix-namespace map in advance so that
the code can understand what namespace a prefix string stands for. See
more details in the demo code.
2 In the classes definition file, please only define the classes within
your ontology. We consider all classes in that file are within your
customized namespace.
3 While reading the properties definition file, we can handle the range
of the properties correctly.
4 we can now generate tables which contains properties information. The
implement is almost the same as the way we generate tables which
contains classes information. Note that now the template DOCX file
contain two tables, the first one is the template of classes
information, the other one is the template of property information.
5 some presentation problem: the width of the cell in the table will
automatically adjust by the content, which may make tables looks
different form each other.